requirements.txt
timm==0.9.16
lightning==2.0.0
pytorch-lightning==2.3.0
albumentations
ttach
numpy
tqdm
opencv-python
scipy
matplotlib
einops
addict
ftfy
regex
searborn
scikit-learn
numpy